Split content saving panel into different files
