+++ utility_section/utility_functions.py
@@ -77,113 +77,3 @@ def build_summary_list(col, mod):
             item.enabled = False
         else:
             item.enabled = True
-
-
-def refresh_shapekeys_ul(self, context):
-    sett = context.scene.HG3D
-    pref = get_prefs()
-    col = context.scene.shapekeys_col
-
-    previously_enabled_items = [i.sk_name for i in col if i.enabled]
-
-    col.clear()
-
-    existing_sks = find_existing_shapekeys(sett.custom_content, pref)
-
-    human = Human.from_existing(context.object)
-    if not human:
-        return
-
-    for sk in human.keys:
-        if sk.name in existing_sks:
-            continue
-
-        item = col.add()
-        item.sk_name = sk.name
-
-        if sk.name in previously_enabled_items:
-            item.enabled = True
-
-        item.on = True if not sk.mute else False
-        if not item.on:
-            item.enabled = False
-
-
-def find_existing_shapekeys(cc_sett, pref):
-    existing_sks = [
-        "Basis",
-    ]
-    if not cc_sett.show_saved_sks:
-        walker = os.walk(str(pref.filepath) + str(Path("/models/shapekeys")))
-        for root, _, filenames in walker:
-            for fn in filenames:
-                if not os.path.splitext(fn)[1] == ".json":
-                    continue
-                with open(os.path.join(root, fn)) as f:
-                    data = json.load(f)
-
-                existing_sks.extend(data)
-    return existing_sks
-
-
-def refresh_hair_ul(self, context):
-    sett = context.scene.HG3D
-    pref = get_prefs()
-    col = context.scene.savehair_col
-
-    previously_enabled_items = [i.ps_name for i in col if i.enabled]
-
-    col.clear()
-
-    hg_rig = sett.content_saving_active_human
-    if not hg_rig:
-        return
-
-    for ps in hg_rig.HG.body_obj.particle_systems:
-        if ps.name.startswith("Eye") and not sett.show_eyesystems:
-            continue
-        item = col.add()
-        item.ps_name = ps.name
-
-        if ps.name in previously_enabled_items:
-            item.enabled = True
-
-
-# TODO if old list, make cloth_types the same again
-def refresh_outfit_ul(self, context):
-    sett = context.scene.HG3D
-    pref = get_prefs()
-    col = context.scene.saveoutfit_col
-
-    previously_enabled_items = [i.obj_name for i in col if i.enabled]
-
-    col.clear()
-
-    hg_rig = sett.content_saving_active_human
-
-    for obj in [
-        o
-        for o in hg_rig.children
-        if o.type == "MESH"
-        and not "hg_body" in o
-        and not "hg_eyes" in o
-        and not "hg_teeth" in o
-    ]:
-
-        item = col.add()
-        item.obj_name = obj.name
-
-        if obj.data.shape_keys:
-            item.cor_sks_present = next(
-                (
-                    True
-                    for sk in obj.data.shape_keys.key_blocks
-                    if sk.name.startswith("cor")
-                ),
-                False,
-            )
-
-        item.weight_paint_present = "spine" in [vg.name for vg in obj.vertex_groups]
-
-        if obj.name in previously_enabled_items:
-            item.enabled = True
